+++ test/v3/asset-data.test.ts
@@ -6,12 +6,12 @@ import {
   SwappableAsset,
   UserFacingERC1155AssetDataSerializedNormalizedSingle,
   UserFacingERC20AssetDataSerialized,
-} from '../src';
+} from '../../src';
 import {
   convertDecodedAssetDataToUserFacingAssets,
   getAssetsFromOrder,
-} from '../src/sdk/v3/pure';
-import { NULL_ADDRESS } from '../src/utils/eth';
+} from '../../src/sdk/v3/pure';
+import { NULL_ADDRESS } from '../../src/utils/eth';
 
 jest.setTimeout(60 * 1000);
