+++ test/v3/forwarder-swap.test.ts
@@ -1,9 +1,7 @@
 import { ethers } from 'ethers';
-import { defaultAbiCoder, hexDataSlice, parseEther } from 'ethers/lib/utils';
-import { NftSwap, SwappableAsset } from '../src';
-import { verifyOrderSignature } from '../src/sdk/v3/pure';
-import { encodeErc20AssetData } from '../src/utils/v3/asset-data';
-import { normalizeOrder } from '../src/utils/v3/order';
+import { parseEther } from 'ethers/lib/utils';
+import { NftSwap, SwappableAsset } from '../../src';
+import { normalizeOrder } from '../../src/utils/v3/order';
 
 jest.setTimeout(60 * 1000);
 
@@ -18,11 +16,12 @@ const MAKER_PRIVATE_KEY =
 // const TAKER_PRIVATE_KEY =
 //   "a8d6d0643c732663bf5221f83df806a59ed54dbd9be02e226b1a11ff4de83de8";
 
-const USDC_TOKEN_ADDRESS_TESTNET = '0xeb8f08a975ab53e34d8a0330e0d34de942c95926';
-const DAI_TOKEN_ADDRESS_TESTNET = '0x6a9865ade2b6207daac49f8bcba9705deb0b0e6d';
+const WMATIC_TOKEN_ADDRESS_TESTNET =
+  '0x9c3c9283d3e44854697cd22d3faa240cfb032889';
+const DAI_TOKEN_ADDRESS_TESTNET = '0x001b3b4d0f3714ca98ba10f6042daebf0b1b7b6f';
 
 const RPC_TESTNET =
-  'https://eth-rinkeby.alchemyapi.io/v2/is1WqyAFM1nNFFx2aCozhTep7IxHVNGo';
+  'https://polygon-mumbai.g.alchemy.com/v2/VMBpFqjMYv2w-MWnc9df92w3R2TpMvSG';
 
 const MAKER_WALLET = new ethers.Wallet(MAKER_PRIVATE_KEY);
 // const TAKER_WALLET = new ethers.Wallet(TAKER_PRIVATE_KEY);
@@ -32,25 +31,45 @@ const PROVIDER = new ethers.providers.StaticJsonRpcProvider(RPC_TESTNET);
 const MAKER_SIGNER = MAKER_WALLET.connect(PROVIDER);
 // const TAKER_PROVIDER = TAKER_WALLET.connect(PROVIDER);
 
-const nftSwapperMaker = new NftSwap(MAKER_SIGNER as any, MAKER_SIGNER, 4);
+const nftSwapperMaker = new NftSwap(MAKER_SIGNER as any, MAKER_SIGNER, 80001);
 // const nftSwapperTaker = new NftSwap(TAKER_PROVIDER as any, 4);
 
-const TAKER_ASSET: SwappableAsset = {
-  type: 'ERC20',
-  tokenAddress: USDC_TOKEN_ADDRESS_TESTNET,
-  amount: '100000', // 1 USDC
-};
 const MAKER_ASSET: SwappableAsset = {
   type: 'ERC20',
   tokenAddress: DAI_TOKEN_ADDRESS_TESTNET,
-  amount: '100000000000000000', // 1 DAI
+  amount: '1000000000', // 1 DAI
+};
+const TAKER_ASSET: SwappableAsset = {
+  type: 'ERC20',
+  tokenAddress: WMATIC_TOKEN_ADDRESS_TESTNET,
+  amount: '1000000000', // 1 WMATIC
 };
 
 describe('NFTSwap', () => {
-  it('swaps 0.1 DAI and 0.1 USDC correctly', async () => {
+  fit('swaps 0.1 DAI and 0.1 WMATIC on mumbai test using forwarder correctly', async () => {
     // NOTE(johnrjj) - Assumes USDC and DAI are already approved w/ the ExchangeProxy
 
     const gasPrice = (await PROVIDER.getGasPrice()).mul(2);
+
+    // const approvalTxMaker = await nftSwapperMaker.approveTokenOrNftByAsset(
+    //   MAKER_ASSET,
+    //   MAKER_WALLET_ADDRESS
+    // );
+    // const approvalTxTaker = await nftSwapperMaker.approveTokenOrNftByAsset(
+    //   TAKER_ASSET,
+    //   MAKER_WALLET_ADDRESS
+    // );
+
+    // const makerApprovalTxReceipt = await approvalTxMaker.wait();
+    // console.log(
+    //   'makerApprovalTxReceipt',
+    //   makerApprovalTxReceipt.transactionHash
+    // );
+    // const takerApprovalTxReceipt = await approvalTxTaker.wait();
+    // console.log(
+    //   'takerApprovalTxReceipt',
+    //   takerApprovalTxReceipt.transactionHash
+    // );
 
     const order = nftSwapperMaker.buildOrder(
       [MAKER_ASSET],
@@ -63,6 +82,11 @@ describe('NFTSwap', () => {
     );
 
     const normalizedOrder = normalizeOrder(order);
+
+    const canOrderBeFilledWithNativeToken =
+      nftSwapperMaker.checkIfOrderCanBeFilledWithNativeToken(order);
+    expect(canOrderBeFilledWithNativeToken).toBe(true);
+
     const signedOrder = await nftSwapperMaker.signOrder(
       normalizedOrder,
       MAKER_WALLET_ADDRESS,
@@ -76,16 +100,23 @@ describe('NFTSwap', () => {
     );
 
     // Uncomment to actually fill order
-    // const tx = await nftSwapperMaker.fillSignedOrder(signedOrder, undefined, {
-    //   gasPrice,
-    //   gasLimit: '500000',
-    //   // HACK(johnnrjj) - Rinkeby still has protocol fees, so we give it a little bit of ETH so its happy.
-    //   value: parseEther('0.01'),
-    // });
+    // const tx = await nftSwapperMaker.fillSignedOrder(
+    //   signedOrder,
+    //   { fillOrderWithNativeTokenInsteadOfWrappedToken: true },
+    //   {
+    //     gasPrice,
+    //     gasLimit: '800000',
+    //     // value: '20000000000000000',
+    //     // value: parseEther('0.000000002'),
+    //   }
+    // );
 
     // const txReceipt = await tx.wait();
     // expect(txReceipt.transactionHash).toBeTruthy();
-    // console.log(`Swapped on Rinkeby (txHAsh: ${txReceipt.transactionIndex})`);
+    // console.log(
+    //   `Forwarder succcess -- Swapped on Mumbai (txHAsh: ${txReceipt.transactionHash})`
+    // );
+    // expect(tx.value.toString()).toBe(signedOrder.takerAssetAmount);
   });
 });
