+++ test/v3/mumbai-swap.test.ts
@@ -1,16 +1,12 @@
 import { ethers } from 'ethers';
 import {
-  AssetProxyId,
-  decodeAssetData,
-  ERC721AssetDataSerialized,
   estimateGasForFillOrder,
-  MultiAssetDataSerializedRecursivelyDecoded,
   NftSwap,
   SupportedChainIds,
   SwappableAsset,
-} from '../src';
-import { DEFAUTLT_GAS_BUFFER_MULTIPLES } from '../src/utils/v3/gas-buffer';
-import { normalizeOrder } from '../src/utils/v3/order';
+} from '../../src';
+import { DEFAUTLT_GAS_BUFFER_MULTIPLES } from '../../src/utils/v3/gas-buffer';
+import { normalizeOrder } from '../../src/utils/v3/order';
 
 jest.setTimeout(60 * 1000);
 
@@ -54,18 +50,6 @@ const MAKER_ASSET: SwappableAsset = {
   amount: '10000000000000000', // 1 DAI
 };
 
-const TAKER_ASSET_NFT_1: SwappableAsset = {
-  type: 'ERC721',
-  tokenAddress: '0xf5de760f2e916647fd766b4ad9e85ff943ce3a2b',
-  tokenId: '4355',
-};
-
-const TAKER_ASSET_NFT_2: SwappableAsset = {
-  type: 'ERC721',
-  tokenAddress: '0xf5de760f2e916647fd766b4ad9e85ff943ce3a2b',
-  tokenId: '4354',
-};
-
 describe('NFTSwap', () => {
   it('swaps 0.1 DAI and 0.1 WMATIC on mumbai test correctly', async () => {
     // NOTE(johnrjj) - Assumes USDC and DAI are already approved w/ the ExchangeProxy
@@ -81,18 +65,6 @@ describe('NFTSwap', () => {
     //   MAKER_WALLET_ADDRESS
     // );
 
-    // const approvalTxTakerNft1 = await nftSwapperMaker.approveTokenOrNftByAsset(
-    //   TAKER_ASSET_NFT_1,
-    //   MAKER_WALLET_ADDRESS
-    // );
-    // await approvalTxTakerNft1.wait()
-
-    // const approvalTxTakerNft2 = await nftSwapperMaker.approveTokenOrNftByAsset(
-    //   TAKER_ASSET_NFT_2,
-    //   MAKER_WALLET_ADDRESS
-    // );
-    // await approvalTxTakerNft2.wait()
-
     // const makerApprovalTxReceipt = await approvalTxMaker.wait();
     // console.log(
     //   'makerApprovalTxReceipt',
@@ -106,7 +78,7 @@ describe('NFTSwap', () => {
 
     const order = nftSwapperMaker.buildOrder(
       [MAKER_ASSET],
-      [TAKER_ASSET, TAKER_ASSET_NFT_1, TAKER_ASSET_NFT_2],
+      [TAKER_ASSET],
       MAKER_WALLET_ADDRESS,
       {
         // Fix dates and salt so we have reproducible tests
@@ -123,57 +95,26 @@ describe('NFTSwap', () => {
 
     const normalizedSignedOrder = normalizeOrder(signedOrder);
 
-    const decodedTakerAssets = decodeAssetData(
-      normalizedSignedOrder.takerAssetData
-    ) as MultiAssetDataSerializedRecursivelyDecoded;
-
     expect(normalizedSignedOrder.makerAddress.toLowerCase()).toBe(
       MAKER_WALLET_ADDRESS.toLowerCase()
     );
 
-    expect(decodedTakerAssets.assetProxyId).toBe(AssetProxyId.MultiAsset);
-    expect(decodedTakerAssets.nestedAssetData.length).toBe(3);
-    expect(decodedTakerAssets.nestedAssetData[0].tokenAddress).toBe(
-      TAKER_ASSET.tokenAddress
-    );
-    expect(decodedTakerAssets.nestedAssetData[1].tokenAddress).toBe(
-      TAKER_ASSET_NFT_1.tokenAddress
-    );
-    expect(
-      (decodedTakerAssets.nestedAssetData[1] as ERC721AssetDataSerialized)
-        .tokenId
-    ).toBe(TAKER_ASSET_NFT_1.tokenId);
-    expect(
-      (decodedTakerAssets.nestedAssetData[2] as ERC721AssetDataSerialized)
-        .tokenId
-    ).toBe(TAKER_ASSET_NFT_2.tokenId);
+    // const estimatedGasLimit = await estimateGasForFillOrder(signedOrder, nftSwapperMaker.exchangeContract);
 
-    const estimatedGasLimit = await estimateGasForFillOrder(
-      signedOrder,
-      nftSwapperMaker.exchangeContract
-    );
-
-    // Uncomment to actually fill order
+    // // Uncomment to actually fill order
     // const tx = await nftSwapperMaker.fillSignedOrder(signedOrder, undefined, {
     //   gasPrice,
     // });
 
-    // const finalGasLimit = tx.gasLimit;
+    // const finalGasLimit = tx.gasLimit
 
-    // const expectedGasLimitWithBufferMultiple = Math.floor(
-    //   estimatedGasLimit.toNumber() *
-    //     DEFAUTLT_GAS_BUFFER_MULTIPLES[SupportedChainIds.PolygonMumbai]
-    // );
+    // const expectedGasLimitWithBufferMultiple = Math.floor(estimatedGasLimit.toNumber() * (DEFAUTLT_GAS_BUFFER_MULTIPLES[SupportedChainIds.PolygonMumbai]))
 
-    // expect(finalGasLimit.toNumber()).toEqual(
-    //   expectedGasLimitWithBufferMultiple
-    // );
+    // expect(finalGasLimit.toNumber()).toEqual(expectedGasLimitWithBufferMultiple)
 
     // const txReceipt = await tx.wait();
     // expect(txReceipt.transactionHash).toBeTruthy();
-    // console.log(
-    //   `Swapped multiasset on Mumbai (txHAsh: ${txReceipt.transactionHash})`
-    // );
+    // console.log(`Swapped on Mumbai (txHAsh: ${txReceipt.transactionHash})`);
   });
 });
