+++ test/v3/order-status.test.ts
@@ -1,7 +1,7 @@
 import { ethers } from 'ethers';
-import { NftSwap, SwappableAsset } from '../src';
-import { OrderStatus } from '../src/sdk/v3/types';
-import { normalizeOrder } from '../src/utils/v3/order';
+import { NftSwap, SwappableAsset } from '../../src';
+import { OrderStatus } from '../../src/sdk/v3/types';
+import { normalizeOrder } from '../../src/utils/v3/order';
 
 jest.setTimeout(60 * 1000);
