+++ test/v3/swap.test.ts
@@ -1,11 +1,9 @@
-import { hexDataLength, hexDataSlice } from '@ethersproject/bytes';
 import { ethers } from 'ethers';
-import { NftSwap, SwappableAsset } from '../src';
-import {
-  signOrderWithEoaWallet,
-  verifyOrderSignature,
-} from '../src/sdk/v3/pure';
-import { normalizeOrder } from '../src/utils/v3/order';
+import { defaultAbiCoder, hexDataSlice, parseEther } from 'ethers/lib/utils';
+import { NftSwap, SwappableAsset } from '../../src';
+import { verifyOrderSignature } from '../../src/sdk/v3/pure';
+import { encodeErc20AssetData } from '../../src/utils/v3/asset-data';
+import { normalizeOrder } from '../../src/utils/v3/order';
 
 jest.setTimeout(60 * 1000);
 
@@ -20,12 +18,11 @@ const MAKER_PRIVATE_KEY =
 // const TAKER_PRIVATE_KEY =
 //   "a8d6d0643c732663bf5221f83df806a59ed54dbd9be02e226b1a11ff4de83de8";
 
-const WMATIC_TOKEN_ADDRESS_TESTNET =
-  '0x9c3c9283d3e44854697cd22d3faa240cfb032889';
-const DAI_TOKEN_ADDRESS_TESTNET = '0x001b3b4d0f3714ca98ba10f6042daebf0b1b7b6f';
+const USDC_TOKEN_ADDRESS_TESTNET = '0xeb8f08a975ab53e34d8a0330e0d34de942c95926';
+const DAI_TOKEN_ADDRESS_TESTNET = '0x6a9865ade2b6207daac49f8bcba9705deb0b0e6d';
 
 const RPC_TESTNET =
-  'https://polygon-mumbai.g.alchemy.com/v2/VMBpFqjMYv2w-MWnc9df92w3R2TpMvSG';
+  'https://eth-rinkeby.alchemyapi.io/v2/is1WqyAFM1nNFFx2aCozhTep7IxHVNGo';
 
 const MAKER_WALLET = new ethers.Wallet(MAKER_PRIVATE_KEY);
 // const TAKER_WALLET = new ethers.Wallet(TAKER_PRIVATE_KEY);
@@ -35,23 +32,25 @@ const PROVIDER = new ethers.providers.StaticJsonRpcProvider(RPC_TESTNET);
 const MAKER_SIGNER = MAKER_WALLET.connect(PROVIDER);
 // const TAKER_PROVIDER = TAKER_WALLET.connect(PROVIDER);
 
-const nftSwapperMaker = new NftSwap(MAKER_SIGNER as any, MAKER_SIGNER, 80001);
+const nftSwapperMaker = new NftSwap(MAKER_SIGNER as any, MAKER_SIGNER, 4);
 // const nftSwapperTaker = new NftSwap(TAKER_PROVIDER as any, 4);
 
 const TAKER_ASSET: SwappableAsset = {
   type: 'ERC20',
-  tokenAddress: WMATIC_TOKEN_ADDRESS_TESTNET,
-  amount: '10000000000000000', // 1 WMATIC
+  tokenAddress: USDC_TOKEN_ADDRESS_TESTNET,
+  amount: '100000', // 1 USDC
 };
 const MAKER_ASSET: SwappableAsset = {
   type: 'ERC20',
   tokenAddress: DAI_TOKEN_ADDRESS_TESTNET,
-  amount: '10000000000000000', // 1 DAI
+  amount: '100000000000000000', // 1 DAI
 };
 
 describe('NFTSwap', () => {
-  it('eoa signatures correctly', async () => {
+  it('swaps 0.1 DAI and 0.1 USDC correctly', async () => {
     // NOTE(johnrjj) - Assumes USDC and DAI are already approved w/ the ExchangeProxy
+
+    const gasPrice = (await PROVIDER.getGasPrice()).mul(2);
 
     const order = nftSwapperMaker.buildOrder(
       [MAKER_ASSET],
@@ -76,25 +75,17 @@ describe('NFTSwap', () => {
       MAKER_WALLET_ADDRESS.toLowerCase()
     );
 
-    const rawSignature = await signOrderWithEoaWallet(
-      order,
-      nftSwapperMaker.signer as any,
-      nftSwapperMaker.chainId,
-      nftSwapperMaker.exchangeContractAddress
-    );
+    // Uncomment to actually fill order
+    // const tx = await nftSwapperMaker.fillSignedOrder(signedOrder, undefined, {
+    //   gasPrice,
+    //   gasLimit: '500000',
+    //   // HACK(johnnrjj) - Rinkeby still has protocol fees, so we give it a little bit of ETH so its happy.
+    //   value: parseEther('0.01'),
+    // });
 
-    const length = hexDataLength(signedOrder.signature);
-    const signatureType = hexDataSlice(signedOrder.signature, length - 1);
-
-    expect(signatureType).toBe('0x02');
-
-    const isValidSignature = await verifyOrderSignature(
-      normalizedSignedOrder,
-      rawSignature,
-      80001,
-      nftSwapperMaker.exchangeContract.address
-    );
-    expect(isValidSignature).toBe(true);
+    // const txReceipt = await tx.wait();
+    // expect(txReceipt.transactionHash).toBeTruthy();
+    // console.log(`Swapped on Rinkeby (txHAsh: ${txReceipt.transactionIndex})`);
   });
 });
